requirements.txt
openai>=1.0.0
youtube-transcript-api>=0.6.0
yt-dlp>=2023.7.6
pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
python-dotenv>=1.0.0
requests>=2.31.0
tenacity>=8.2.0
argparse-color-formatter>=1.2.2
colorama>=0.4.6
tqdm>=4.65.0
imagehash>=4.3.1
numpy>=1.24.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0